Event segment validation functionality
- Every segment distance check with above and below value
- Every segment goal time check with above and below value
- Combined segment distance exceeds total distance
- Combined segment ETA exceeds total goal
- Combined segment distance is below total distance
- Combined segment ETA is below total goal

+++ Modules/Dashboard/Home/NewRun/Model /RunSegment.swift
@@ -10,7 +10,7 @@ import Foundation
 // MARK: - Segment Model
 struct RunSegment: Identifiable {
     let id: Int
-    var distance: Double
+    var distance: Float
 	var goalHours: Int
     var goalMinutes: Int
     var goalSeconds: Int
